src/github_preparation/requirements.txt
gql
python-dotenv
requests
requests_toolbelt
openai
argparse